fixed depreciation error for np.float

+++ End_to_End_Autonomous_Driving/notebooks/data.py
@@ -425,7 +425,7 @@ def draw_target_point(target_point, color = (255, 255, 255)):
     point = np.clip(point, 0, 256)
     cv2.circle(image, tuple(point), radius=5, color=color, thickness=3)
     image = image.reshape(1, 256, 256)
-    return image.astype(np.float) / 255.
+    return image.astype(np.float32) / 255.
 
 
 def decode_pil_to_npy(img):
